app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
neo4j==5.20.0
httpx==0.27.2
requests==2.32.3
openai>=1.40.0
groq>=0.9.0
tqdm==4.66.4
orjson==3.10.7
sentence-transformers==2.7.0